+++ eno-core/src/main/java/fr/insee/eno/core/DDIToLunatic.java
@@ -33,7 +33,7 @@ public class DDIToLunatic implements InToOut<Questionnaire> {
      * @throws DDIParsingException if the input stream given cannot be parsed to a DDI object.
      * @deprecated use other transform method.
      */
-    @Deprecated(since = "3.23.0")
+    @Deprecated(since = "3.33.0")
     public Questionnaire transform(InputStream ddiInputStream, EnoParameters enoParameters)
             throws DDIParsingException {
         ddiQuestionnaire = DDIDeserializer.deserialize(ddiInputStream);