+++ eno-core/src/main/java/fr/insee/eno/core/InToEno.java
@@ -11,8 +11,9 @@ public interface InToEno {
     /**
      * @deprecated use other transform method.
      */
-    @Deprecated(since = "3.23.0")
+    @Deprecated(since = "3.33.0")
     EnoQuestionnaire transform(InputStream inputStream, EnoParameters enoParameters) throws ParsingException;
+
     EnoQuestionnaire transform(EnoParameters enoParameters);
 
 }