+++ eno-core/src/main/java/fr/insee/eno/core/PoguesToLunatic.java
@@ -32,7 +32,7 @@ public class PoguesToLunatic implements InToOut<Questionnaire> {
      * @throws PoguesDeserializationException if the input stream given cannot be parsed to a Pogues questionnaire.
      * @deprecated use other transform method.
      */
-    @Deprecated(since = "3.23.0")
+    @Deprecated(since = "3.33.0")
     public Questionnaire transform(InputStream poguesInputStream, EnoParameters enoParameters)
             throws PoguesDeserializationException {
         poguesQuestionnaire = PoguesDeserializer.deserialize(poguesInputStream);